Move multilevel tests to a dedicated file
